refactor(agent): isolate VM recovery execution
