+++ packages/agent/test/_helpers/vm-recovery-host.ts
@@ -9,7 +9,7 @@ import type { VmReconcileRotationRecord } from '../../src/dkg-agent-types.js';
 import { DKGAgent } from '../../src/index.js';
 import type {
   VmRecoveryUalDisposition,
-} from '../../src/vm-recovery-microbatch.js';
+} from '../../src/vm-recovery-provider-policy.js';
 import type { VmRecoveryChainFootprint } from '../../src/vm-recovery-types.js';
 
 interface TestPeerId {
